// ui/src/lib/api/schema.ts
export interface components {
    schemas: {
// ...
        NetTerms: {
            /** Mv Base */
            mv_base?: number | null;
            /** Invested */
            invested?: number | null;
            /** Partial */
            partial?: string | null;
        };
// ...
        WindowBaseResponse: {
            /** By Period */
            by_period: {
                [key: string]: {
                    [key: string]: number | null;
                } | null;
            };
            /**
             * Market Value
             * @default {}
             */
            market_value: {
                [key: string]: {
                    [key: string]: number | null;
                } | null;
            };
            /**
             * Net
             * @default {}
             */
            net: {
                [key: string]: {
                    [key: string]: components["schemas"]["NetTerms"];
                } | null;
            };
            /**
             * Base Date
             * @default {}
             */
            base_date: {
                [key: string]: string | null;
            };
            /**
             * Unreadable
             * @default []
             */
            unreadable: string[];
            /** Error */
            error?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}